Move PreferPHPUnitSelfCallRector, CreateMockToAnonymousClassRector from incorrect /src path to /rules
