+++ rules-tests/CodeQuality/Rector/ClassMethod/CreateMockToAnonymousClassRector/Fixture/mocked_more_methods.php.inc
@@ -1,9 +1,9 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector\Fixture;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\ClassMethod\CreateMockToAnonymousClassRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
-use Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector\Source\ToBeFixturedClass;
+use Rector\PHPUnit\Tests\CodeQuality\Rector\ClassMethod\CreateMockToAnonymousClassRector\Source\ToBeFixturedClass;
 
 final class MockedMoreMethods extends TestCase
 {
@@ -23,10 +23,10 @@ final class MockedMoreMethods extends TestCase
 -----
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector\Fixture;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\ClassMethod\CreateMockToAnonymousClassRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
-use Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector\Source\ToBeFixturedClass;
+use Rector\PHPUnit\Tests\CodeQuality\Rector\ClassMethod\CreateMockToAnonymousClassRector\Source\ToBeFixturedClass;
 
 final class MockedMoreMethods extends TestCase
 {
