+++ rules-tests/CodeQuality/Rector/ClassMethod/CreateMockToAnonymousClassRector/Fixture/skip_with_args.php.inc
@@ -1,9 +1,9 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector\Fixture;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\ClassMethod\CreateMockToAnonymousClassRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
-use Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector\Source\ToBeFixturedClass;
+use Rector\PHPUnit\Tests\CodeQuality\Rector\ClassMethod\CreateMockToAnonymousClassRector\Source\ToBeFixturedClass;
 
 final class SkipWithArgs extends TestCase
 {
