+++ rules-tests/CodeQuality/Rector/ClassMethod/CreateMockToAnonymousClassRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\Rector\ClassMethod\CreateMockToAnonymousClassRector;
+use Rector\PHPUnit\CodeQuality\Rector\ClassMethod\CreateMockToAnonymousClassRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(CreateMockToAnonymousClassRector::class);
