+++ rules-tests/CodeQuality/Rector/Class_/PreferPHPUnitSelfCallRector/Fixture/skip_this_local.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\PHPUnit\Tests\Rector\Class_\PreferPHPUnitSelfCallRector\Fixture;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\PreferPHPUnitSelfCallRector\Fixture;
 
 use PHPUnit\Framework\TestCase;
