+++ rules-tests/CodeQuality/Rector/Class_/PreferPHPUnitSelfCallRector/PreferPHPUnitSelfCallRectorTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector;
+namespace Rector\PHPUnit\Tests\CodeQuality\Rector\Class_\PreferPHPUnitSelfCallRector;
 
 use Iterator;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-final class CreateMockToAnonymousClassRectorTest extends AbstractRectorTestCase
+final class PreferPHPUnitSelfCallRectorTest extends AbstractRectorTestCase
 {
     #[DataProvider('provideData')]
     public function test(string $filePath): void
