+++ rules-tests/CodeQuality/Rector/Class_/PreferPHPUnitSelfCallRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\Rector\Class_\PreferPHPUnitSelfCallRector;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferPHPUnitSelfCallRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(PreferPHPUnitSelfCallRector::class);
