--- tests/Rector/ClassMethod/CreateMockToAnonymousClassRector/Source/ToBeFixturedClass.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Rector\PHPUnit\Tests\Rector\ClassMethod\CreateMockToAnonymousClassRector\Source;
-
-final class ToBeFixturedClass
-{
-
-}
